added in 'more' and did work on local web views... not currently 100% functional
