+++ VdubApp/home.html
@@ -1,0 +1,72 @@
+<!doctype html>
+
+<html lang="en">
+    <head>
+        <meta charset="utf-8">
+            
+            <title>Eatery Hours</title>
+            <style>
+                h1 {
+                    padding:0;
+                    margin:0;
+                    font-size:16px;
+                    font-family:sans-serif;
+                }
+            p {
+                padding:0;
+                margin:0;
+                font-size:12px;
+                font-family:sans-serif;
+            }
+            </style>
+            </head>
+    <body>
+        <p>~ hours may vary seasonally ~</p>
+        
+        <h1>Sharpe Refectory ("Ratty")</h1>
+        <p>Mon-Sat 7:30a-7:30p</p><br/>
+        <p>Sunday 10:30a-8p</p><br/>
+        
+        <h1>Verney-Woolley ("V-Dub")</h1>
+        <p>Weekdays only</p><br/>
+        <p>Breakfast 7:30a-9:30a</p><br/>
+        <p>Continental 9:30-11a</p><br/>
+        <p>Lunch 11a-2p</p><br/>
+        <p>Dinner 4:30p-7:30p</p><br/>
+        
+        <h1>Andrews Commons</h1>
+        <p>Everday 11a-2a</p><br/>
+        
+        <h1>Blue Room Cafe</h1>
+        <p>Weekdays 7:30a-9p</p><br/>
+        <p>Weekends 9a-5p</p><br/>
+        
+        <h1>Campus Market</h1>
+        <p>Weekdays 10a-11p</p><br/>
+        <p>Weekends 2p-11p</p><br/>
+        
+        <h1>Faculty Club</h1>
+        <p>Weekdays Lunch 11:30a-2p</p><br/>
+        <p>Weds-Sat Dinner 5p-8p</p><br/>
+        
+        <h1>Friedman Cafe</h1>
+        <p>Mon-Thurs 8:30a-1a</p><br/>
+        <p>Friday 8:30a-6p</p><br/>
+        <p>Saturday 10a-10p</p><br/>
+        <p>Sunday 10a-1a</p><br/>
+        
+        <h1>Ivy Room</h1>
+        <p>Weekdays Lunch 11:30a-1:45p</p><br/>
+        <p>Sun-Thurs Night 7:45p-12a</p><br/>
+        
+        <h1>Josiah's</h1>
+        <p>Everday 6p-2a</p><br/>
+        
+        <h1>Little Jo's</h1>
+        <p>Everyday 6p-1:30a</p><br/>
+        
+        <h1>Poppy's</h1>
+        <p>Weekdays 8a-8p</p><br/>
+        <p>Weekends 10a-8p</p><br/>
+    </body>
+</html>
